third_party/blink/web_tests/clipboard/async-clipboard/writetext-exception-stack.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

test(() => {
  navigator.clipboard.writeText("")
    .then(() => assert_unreached("unexpected success result"))
    .catch((err) => assert_true(!!err.stack, "should have a stack"));
});

</script>
